[PK-209] Modification of Celesta maven-plugin and celesta-unit checkstyle: VisibilityModifier.packageAllowed -> true refactoring: ru.curs.celesta.plugin -> ru.curs.celesta.plugin.maven
