+++ celesta-maven-plugin/src/main/java/ru/curs/celesta/plugin/maven/GenTestScoreResourcesMojo.java
@@ -1,4 +1,4 @@
-package ru.curs.celesta.plugin;
+package ru.curs.celesta.plugin.maven;
 
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
