+++ celesta-maven-plugin/src/main/java/ru/curs/celesta/plugin/maven/ScoreProperties.java
@@ -1,4 +1,4 @@
-package ru.curs.celesta.plugin;
+package ru.curs.celesta.plugin.maven;
 
 import org.apache.maven.plugins.annotations.Parameter;
