+++ celesta-maven-plugin/src/test/java/ru/curs/celesta/plugin/maven/AbstractCelestaMojoTestCase.java
@@ -1,4 +1,4 @@
-package ru.curs.celesta.plugin;
+package ru.curs.celesta.plugin.maven;
 
 import java.io.File;
 import java.io.IOException;
