+++ celesta-maven-plugin/src/test/java/ru/curs/celesta/plugin/maven/CelestaMavenPluginStub.java
@@ -1,4 +1,4 @@
-package ru.curs.celesta.plugin;
+package ru.curs.celesta.plugin.maven;
 
 import org.apache.maven.model.Build;
 import org.apache.maven.model.Model;
