+++ celesta-maven-plugin/src/test/resources/pom.xml
@@ -14,7 +14,7 @@
                 <artifactId>celesta-maven-plugin</artifactId>
                 <version>${project.version}</version>
                 <configuration>
-                    <project implementation="ru.curs.celesta.plugin.CelestaMavenPluginStub"/>
+                    <project implementation="ru.curs.celesta.plugin.maven.CelestaMavenPluginStub"/>
                 </configuration>
             </plugin>
         </plugins>
