+++ celesta-maven-plugin/src/test/resources/pom_badScore.xml
@@ -24,7 +24,7 @@
                             <path>src/test/score/resources/badScore</path>
                         </score>
                     </testScores>
-                    <project implementation="ru.curs.celesta.plugin.CelestaMavenPluginStub"/>
+                    <project implementation="ru.curs.celesta.plugin.maven.CelestaMavenPluginStub"/>
                 </configuration>
             </plugin>
         </plugins>
